Conflicts:
	lib/scip-git

+++ src/dec_borderheur.c
@@ -577,7 +577,7 @@ static SCIP_RETCODE buildTransformedProblem(
    /* go through all of the constraints */
    for( i = 0; i < nconss; i++ )
    {
-      long int consblock = -1;
+      int consblock = -1;
       int ncurvars;
       SCIP_VAR **curvars;
 
@@ -596,7 +596,7 @@ static SCIP_RETCODE buildTransformedProblem(
       for( j = 0; j < ncurvars; j++ )
       {
          SCIP_VAR* var;
-         long int varblock;
+         int varblock;
          assert(curvars != NULL);
 
          if( !SCIPisVarRelevant(curvars[j]) )
@@ -631,11 +631,11 @@ static SCIP_RETCODE buildTransformedProblem(
 
             /* finally set the hashmap image */
             assert(!SCIPhashmapExists(vartoblock, var));
-            SCIP_CALL( SCIPhashmapInsert(vartoblock, var, (void*)varblock) );
+            SCIP_CALL( SCIPhashmapInsert(vartoblock, var, (void*) (size_t) varblock) );
          }
          else
          {
-            varblock = (long int)SCIPhashmapGetImage(vartoblock, var);
+            varblock = (int)(size_t)SCIPhashmapGetImage(vartoblock, var);
             assert(varblock == detectordata->varpart[SCIPvarGetProbindex(var)] ||  detectordata->varpart[SCIPvarGetProbindex(var)] == -2);
          }
 
@@ -696,7 +696,7 @@ static SCIP_RETCODE buildTransformedProblem(
       {
          subscipconss[consblock][nsubscipconss[consblock]] = conss[i];
          assert(!SCIPhashmapExists(constoblock, conss[i]));
-         SCIP_CALL( SCIPhashmapInsert(constoblock, conss[i], (void*)consblock) );
+         SCIP_CALL( SCIPhashmapInsert(constoblock, conss[i], (void*) (size_t) consblock) );
          ++(nsubscipconss[consblock]);
       }
